Move Flask project to different directory
