+++ _src/resources/pre_config.py
@@ -2,8 +2,8 @@ import requests
 import mongoengine as me
 from flask import request
 from flask_restful import Resource
-from src.model.pre_config import PreConfig
-from src.resources.utils import simple_error_response
+from _src.model.pre_config import PreConfig
+from _src.resources.utils import simple_error_response
 
 
 DUPLICATED_PRE_CONFIG_NAME_MSG = "The pre config name is already in use"
