+++ tests/conftest.py
@@ -1,7 +1,7 @@
 import pytest
-from src.app import create_app
-from src.model.pre_config import PreConfig
-from src.model.metrics_component_tree import MetricsComponentTree
+from _src.app import create_app
+from _src.model.pre_config import PreConfig
+from _src.model.metrics_component_tree import MetricsComponentTree
 
 
 @pytest.fixture()
